+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.0.5
 
 - Optional `BETTY_CACHE_FLUSHER` callback now passed list of string URLS instead of individual strings. This allows for more efficient callback batching logic.
+- Added reference cache flusher `betty.contrib.cacheflush.cachemaster`
 
 ## Version 2.0.4